Put LVMCode and LVMGenerator in their own file
